requirements.txt
allure-behave==2.4.1
allure-python-commons==2.4.1
attrs==18.1.0
behave==1.2.6
certifi==2018.4.16
chardet==3.0.4
idna==2.7
json2html==1.2.1
numpy==1.14.5
pandas==0.23.1
parse==1.8.4
parse-type==0.4.2
pluggy==0.6.0
python-dateutil==2.7.3
pytz==2018.4
requests==2.19.1
six==1.11.0
urllib3==1.23
